refactor: move component files to logical places
Remove the use of __test__ directory.
