+++ apps/ui/modules/reservable.test.ts
@@ -12,7 +12,7 @@ import {
   generateReservableMap,
   isRangeReservable,
   isStartTimeValid,
-} from "../reservable";
+} from "./reservable";
 import {
   BlockingReservationFieldsFragment,
   type IsReservableFieldsFragment,
