+++ packages/common/src/common/util.test.ts
@@ -1,9 +1,9 @@
 import {
   type ReservationNode,
   ReservationStartInterval,
-} from "../../../gql/gql-types";
-import { getEventBuffers } from "../../calendar/util";
-import { getIntervalMinutes } from "../../conversion";
+} from "./../../gql/gql-types";
+import { getEventBuffers } from "./../calendar/util";
+import { getIntervalMinutes } from "./../conversion";
 import { vi, describe, test, expect } from "vitest";
 
 const tfunction = (str: string, options: { count: number }) => {
